requirements.txt
streamlit>=1.20
geopandas>=0.14
rasterio>=1.4
numpy>=1.24
pandas>=2.0
plotly>=5.16
shapely>=2.0
pyproj>=3.5
pydeck>=0.8
folium>=0.14
matplotlib>=3.5
Pillow>=9.0
